refactor: remove dead files and move feature-specific utils to feature folders
- Remove unused files: Personal.tsx, HoverContainer.tsx, useNetworkStatus.ts
- Move makersList.ts to features/makers/ (only used by makers feature)
- Move ReviewWriteTabs.tsx to features/write-reviews/types/ (feature-specific)
- Move checkOverlap.ts to features/timetable/utils/ (timetable-specific)
