+++ app/routes/makers.tsx
@@ -5,7 +5,7 @@ import styled from "@emotion/styled"
 import FlexWrapper from "@/common/primitives/FlexWrapper"
 import MemberSection from "@/features/makers/sections/MemberSection"
 import ProjectSection from "@/features/makers/sections/ProjectSection"
-import { MakersList } from "@/utils/makersList"
+import { MakersList } from "@/features/makers/makersList"
 
 const MakersWrapper = styled(FlexWrapper)`
     flex: 1 0 auto;
