--- app/utils/useNetworkStatus.ts
@@ -1,22 +1,0 @@
-import { useEffect, useState } from "react"
-
-export function useNetworkStatus() {
-    const [isOnline, setIsOnline] = useState(() =>
-        typeof navigator !== "undefined" ? navigator.onLine : true,
-    )
-
-    useEffect(() => {
-        const handleOnline = () => setIsOnline(true)
-        const handleOffline = () => setIsOnline(false)
-
-        window.addEventListener("online", handleOnline)
-        window.addEventListener("offline", handleOffline)
-
-        return () => {
-            window.removeEventListener("online", handleOnline)
-            window.removeEventListener("offline", handleOffline)
-        }
-    }, [])
-
-    return isOnline
-}
